+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Added WP-CLI 'chubes codebase tree' command to display taxonomy hierarchy
+- Improved logging in 'chubes codebase ensure' command
+
 ## [0.5.4] - 2026-01-23
 
 ### Added